Fix: match printf semantics for %08X on negative values
std::format preserves the sign of a signed argument where printf's
%X treats it as unsigned. Cast to unsigned int so the conversion is
equivalent for every value, not just the non-negative ones.

+++ src/SexyAppFramework/misc/Buffer.cpp
@@ -110,7 +110,7 @@ std::string Buffer::ToWebString() const
 	int anOldReadBitPos = mReadBitPos;
 	mReadBitPos = 0;
 
-	aString += std::format("{:08X}", aSizeBits);
+	aString += std::format("{:08X}", static_cast<unsigned int>(aSizeBits));
 
 	int aNumChars = (aSizeBits + 5) / 6;
 	for (int aCharNum = 0; aCharNum < aNumChars; aCharNum++)
